requirements.txt
Flask==2.3.3
Werkzeug==3.1.3
pyodbc==4.0.39
python-dotenv==1.0.0
bcrypt==4.0.1
flake8==6.1.0
Flask-Mail==0.10.0
requests==2.31.0
pyotp==2.8.0
qrcode[pil]==7.4.2
pytest==8.2.2
selenium==4.21.0
Flask-SQLAlchemy==3.1.1
flake8==6.1.0
flake8-sarif==1.0.15
bandit==1.8.6
Flask-Session==0.8.0